requirements/requirements-essential.txt
setuptools
django-dirtyfields==1.4.1
hashids
html2text
trml2pdf==0.5.0

# PostgreSQL python client
psycopg2-binary==2.7.7

# Asynchronous Task Queue
celery==4.3.0

# Djangoplicity actions
git+https://@github.com/djangoplicity/djangoplicity-actions@develop

# Djangoplicity
git+https://@github.com/djangoplicity/djangoplicity.git@develop

# Djangoplicity newsletters and mailinglist
git+https://github.com/djangoplicity/djangoplicity-newsletters.git@develop

# RSA > 4.0 does not support Python 3 (This is a django dependency)
rsa==4.0

# Faker
Faker==3.0.1

# Asset packaging library for Django
django-pipeline==1.6.14

# JSMin latest version doesn't support Python 2
jsmin==2.2.2

# The Django Debug Toolbar is a configurable set of panels that display various debug information
django-debug-toolbar==1.11

# Library to create spreadsheet files compatible with MS Excel
xlwt

# Django DRY forms
django-crispy-forms==1.8.1